Refactor: organize query-related files
