--- Justfile
@@ -1,57 +1,0 @@
-# バックエンドのディレクトリ
-BACKEND_DIR := 'backend'
-
-# フロントエンドのディレクトリ
-FRONTEND_DIR := 'frontend'
-
-# フロントエンドのWebディレクトリ
-FRONTEND_WEB_DIR := 'frontend/web'
-
-# フロントエンドのAppディレクトリ
-FRONTEND_APP_DIR := 'frontend/app'
-
-# バックエンドの起動
-backend:
-    cd {{BACKEND_DIR}} && php artisan serve
-
-backend-run *args:
-    cd {{BACKEND_DIR}} && {{ args }}
-
-pail:
-    cd {{BACKEND_DIR}} && php artisan pail --timeout=600
-
-queue:
-    cd {{BACKEND_DIR}} && php artisan queue:listen --tries=1 --timeout=600
-
-ziggy:
-    cd {{BACKEND_DIR}} && php artisan ziggy:generate
-
-route:
-    cd {{BACKEND_DIR}} && php artisan route:list
-
-fresh:
-    cd {{BACKEND_DIR}} && php artisan migrate:fresh --seed
-
-# フロントエンドの起動
-web:
-    cd {{FRONTEND_WEB_DIR}} && pnpm run dev
-
-storybook:
-    cd {{FRONTEND_WEB_DIR}} && pnpm run storybook --no-open
-
-web-run *args:
-    cd {{FRONTEND_WEB_DIR}} && {{ args }}
-
-app:
-    cd {{FRONTEND_APP_DIR}} && pnpm tauri dev
-
-# バックエンドとフロントエンドの同時起動
-all:
-    #!/bin/bash -eu
-    just ziggy
-    (stdbuf -oL just backend | sed "s/^/$(printf '\033[32m[BACKEND]\033[0m ')/") &
-    (stdbuf -oL just pail | sed "s/^/$(printf '\033[31m[PAIL]\033[0m ')/") &
-    (stdbuf -oL just queue | sed "s/^/$(printf '\033[35m[QUEUE]\033[0m ')/") &
-    (stdbuf -oL just web | sed "s/^/$(printf '\033[34m[FRONTEND]\033[0m ')/") &
-    trap 'kill $(jobs -pr)' EXIT
-    wait
