+++ frontend/core/src/routes/auth/route.tsx
@@ -1,5 +1,5 @@
 import Authenticated from "@/scripts/layouts/AuthenticatedLayout";
-import { requireAuth } from "@lib/auth";
+import { requireAuth } from "@/scripts/queries/auth";
 import { Outlet, createFileRoute } from "@tanstack/react-router";
 
 export const Route = createFileRoute("/auth")({
