+++ frontend/core/src/scripts/pages/profile/partials/UpdatePasswordForm.tsx
@@ -1,9 +1,9 @@
+import { callApi } from "@/scripts/queries/apiClient";
 import type { Auth } from "@/scripts/types";
 import InputLabel from "@components/InputLabel";
 import PrimaryButton from "@components/PrimaryButton";
 import TextInput from "@components/TextInput";
 import { Transition } from "@headlessui/react";
-import { callApi } from "@lib/apiClient";
 import { useForm } from "@tanstack/react-form";
 import { useMutation, useQueryClient } from "@tanstack/react-query";
 import { useRef, useState } from "react";
