+++ frontend/core/src/scripts/queries/auth.ts
@@ -1,8 +1,8 @@
 import type { User } from "@/scripts/types";
-import { callApi } from "@lib/apiClient";
-import { queryClient } from "@lib/queryClient";
 import { redirect } from "@tanstack/react-router";
 import { route } from "ziggy-js";
+import { callApi } from "./apiClient";
+import { queryClient } from "./queryClient";
 
 export const fetchAuth = async (): Promise<{ user: User }> => {
   const user = await callApi("get", route("get-user"));
